Add parse table for error
Add the parse table for error, and update the non terminal code tag in
order to resolve the conflict that arises in case a tags has an
attribute and a child tag with the same name. In this case the string
'_attribute' is appended to the key in the final dictionary.